Further reduce requires, particularly for Windows. Output processing moved from energy_rating_index.rb to design.rb.
